Fix OAM DMA restart logic
Writing to the DMA register during a DMA transfer should restart the transfer.
For fresh DMA transfers, OAM is not accessible while in the "active" state. But
during restarted DMA transfers, in addition to the "active" state, OAM is not
accessible in the "requested" or "starting" states (anything other than "idle").

This fixes the mooneye/oam_dma_restart.gb test.