Correctly calculate final color/shade in pixel mixing logic
The previous implementation only calculated a pixel's final color if a
sprite was present on the FIFO, causing non-sprite areas to be drawn
as Color 0.

This commit restructures the logic to always calculate the background
pixel's shade first, and then conditionally overwrites it if a
non-transparent sprite with the correct priority is present.