Fix sprite wrapping and optimize FIFO clearing
Fixed a visual bug where sprites on the far right edge of the screen wrapped
around to the left side of the next scanline. This occurred because the Sprite
FIFO was not being cleared between lines. An example of this occuring was the
star sprites in Kirby's Dream Land.

Also changed FIFO clearing logic from setting `nil` to reslicing `[:0]`. This
ensures sprite data is correctly reset while reusing the underlying memory,
reducing garbage collection pressure during the scanline loop.